Create a separate polyfills/smoothscroll.ts file and have it imported into polyfills.ts
